\name{sonicLength-package}
\alias{sonicLength-package}
\alias{sonicLength}
\docType{package}
\title{
  Calibration of Number of Length Variants
}
\description{
  Estimate the number of fragments produced by sonication
  of integration sites from tallies of the number of
  distinct lengths.
}
\details{
\tabular{ll}{
Package: \tab sonicLength\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-24\cr
License: \tab see file LICENSE distributed with this software\cr
LazyLoad: \tab yes\cr
}

This package has the necessary utilities to provide a function that will
do calibration of the number of distinct lengths (for an integration
site recovered via sonication) to estimate the number of distinct
sonication fragments that underly the number of lengths.

}
\author{
Charles C. Berry

Maintainer: <ccberry@users.r-forge.r-project.org>
}
